🌧 :: (#8) post폴더 생성 후 postList와 postCard를 분리함
